Release of 5.0 (#311)
* Closes #287 (#288)

* Update README.md

* fixed stupid typo i made (#290)

* Issue 293 command is shared amongst all new controls (#294)

* Disabled possibly transparent background colors, fixes #291

* restored default control functionality - closes #292

* closes #280

* fixed commands being backwards (#296)

* fixed commands being referenced between commands (#297)

* Issue 299 not sending numpad commands correctly (#300)

* Fixes numpad numeric entry

* Updated SC screen templates

* Fixed shields for SC large tablet

* added null check to background path when changing background from color to image

* fixed display not updating after deleting a control when the background image is set

* removed secondary text field for toggles (#306)

* removed secondary text field for toggles

* Fixes in place to allow using the same filename but a different file

* warning only shows on A12 or higher (#308)

* fixed threading issue importing multiple screens at once (#310)

+++ gic_flutter/lib/src/views/screenEditor/backgroundDialog.dart
@@ -53,16 +53,20 @@ class _BackgroundDialogState extends State<BackgroundDialog> {
   }
 
   void _pickBackgroundImage() async {
+    await FilePicker.platform.clearTemporaryFiles();
     FilePickerResult result = await FilePicker.platform.pickFiles(
       type: FileType.custom,
       allowedExtensions: ['jpg', 'png', 'gif'],
     );
     if (result != null) {
+      imageCache.clear();
+      imageCache.clearLiveImages();
+
       PlatformFile file = result.files.first;
       Directory dest = await getApplicationDocumentsDirectory();
       String destPath = path.join(
           dest.path, "screens", screenViewModel.screenId.toString(), path.basename(file.path));
-      if (File(screenViewModel.backgroundPath).existsSync())
+      if (screenViewModel.backgroundPath != null && File(screenViewModel.backgroundPath).existsSync())
         File(screenViewModel.backgroundPath).deleteSync();
       File newFile = File(file.path).copySync(destPath);
       setState(() {
